+++ server/routes/polls/update.route.js
@@ -1,9 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const Poll = require("../../models/polls");
-const multer = require("multer");
-const aws = require("aws-sdk");
-const multerS3 = require("multer-s3");
+const upload = require("./utils");
 
 // Route to update the title of the poll
 router.post("/", (req, res) => {
@@ -15,25 +13,6 @@ router.post("/", (req, res) => {
 	});
 
 	return res.status(200).json({ status: "ok" });
-});
-
-const s3 = new aws.S3({
-	secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-	accessKeyId: process.env.AWS_ACCESS_KEY_ID,
-});
-
-// Middleware to upload the image
-const upload = multer({
-	storage: multerS3({
-		s3: s3,
-		bucket: "hatchways-social-app",
-		metadata: (req, file, cb) => {
-			cb(null, { fieldName: file.fieldname });
-		},
-		key: function (req, file, cb) {
-			cb(null, Date.now().toString() + "-" + file.originalname);
-		},
-	}),
 });
 
 // Router to update the image
